requirements.txt
streamlit
confluent-kafka
psycopg2-binary
requests
python-dotenv
